requirements.txt
Django==2.1.5
djangorestframework==3.9.1
future==0.17.1
pytz==2018.9
